requirements.txt
allure-pytest==2.12.0
allure-python-commons==2.12.0
async-generator==1.10
attrs==22.2.0
autokey==0.95.10
autopep8==1.7.0
bcrypt==3.2.0
beautifulsoup4==4.10.0
blinker==1.4
cachetools==5.0.0
certifi==2022.12.7
cffi==1.15.1
chardet==4.0.0
click==8.0.3
colorama==0.4.4
command-not-found==0.3
cryptography==3.4.8
cupshelpers==1.0
dbus-python==1.2.18
defer==1.0.6
distlib==0.3.6
distro==1.7.0
distro-info===1.1build1
dnspython==2.1.0
duplicity==0.8.21
enum34==1.1.10
evdev==1.4.0
exceptiongroup==1.1.0
fasteners==0.14.1
filelock==3.9.0
future==0.18.2
google-api-python-client==1.7.11
google-auth==1.5.1
google-auth-httplib2==0.1.0
gpg===1.16.0-unknown
greenlet==1.1.2
h11==0.14.0
html5lib==1.1
httplib2==0.20.2
idna==3.3
importlib-metadata==4.6.4
iniconfig==2.0.0
jeepney==0.7.1
keyring==23.5.0
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lockfile==0.12.2
lxml==4.8.0
macaroonbakery==1.3.1
Magnus==1.0.3
Markdown==3.3.6
meson==0.61.2
monotonic==1.6
more-itertools==8.10.0
MouseInfo==0.1.3
namedlist==1.8
netifaces==0.11.0
notify2==0.3
oauth2client==4.1.3
oauthlib==3.2.0
olefile==0.46
onboard==1.4.1
outcome==1.2.0
packaging==23.0
paramiko==2.9.3
pexpect==4.8.0
pibootctl==0.5.2
Pillow==9.0.1
platformdirs==2.6.2
pluggy==1.0.0
protobuf==3.12.4
psutil==5.9.0
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.1
PyAutoGUI==0.9.53
pycairo==1.20.1
pycodestyle==2.9.1
pycparser==2.21
pycups==2.0.1
PyDrive==1.3.1
PyGetWindow==0.0.9
Pygments==2.11.2
PyGObject==3.42.1
pyinotify==0.9.6
PyJWT==2.3.0
pymacaroons==0.13.0
PyMsgBox==1.0.9
PyNaCl==1.5.0
pyOpenSSL==21.0.0
pyparsing==2.4.7
pyperclip==1.8.2
PyRect==0.2.0
pyRFC3339==1.1
PyScreeze==0.1.28
PySocks==1.7.1
pytest==7.2.1
python-apt==2.4.0+ubuntu1
python-dateutil==2.8.1
python-debian===0.1.43ubuntu1
python-dotenv==0.21.0
python-magic==0.4.24
python-xapp==2.2.1
python-xlib==0.29
python3-xlib==0.15
pytweening==1.0.4
pytz==2022.1
pyudev==0.22.0
PyYAML==5.4.1
requests==2.25.1
requests-toolbelt==0.9.1
rsa==4.8
SecretStorage==3.3.1
selenium==4.7.2
setproctitle==1.2.2
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.1
SQLAlchemy==1.4.31
systemd-python==234
toml==0.10.2
tomli==2.0.1
tqdm==4.64.1
trio==0.22.0
trio-websocket==0.9.2
ubuntu-advantage-tools==8001
ubuntu-cleaner==1.1.4
ubuntu-drivers-common==0.0.0
ufw==0.36.1
unattended-upgrades==0.1
undetected-chromedriver==3.4.6
uritemplate==3.0.1
urllib3==1.26.5
virtualenv==20.17.1
wadllib==1.3.6
webdriver-manager==3.8.5
webencodings==0.5.1
websockets==10.4
wsproto==1.2.0
xkit==0.0.0
xlrd==1.2.0
zipp==1.0.0
